pyproject.toml
[project]
name = "zeta-dataset"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.49.1",
    "datasets>=4.3.0",
    "jupyter>=1.1.1",
    "openai>=2.16.0",
    "peft>=0.18.1",
    "python-dotenv>=1.2.1",
    "sentencepiece>=0.2.1",
    "torch>=2.10.0",
    "transformers>=4.57.6",
    "trl>=0.24.0",
    "unsloth>=2026.1.4",
]
